Add files via upload
I have a lot of files that are several page pdfs that I need to split into single page pdfs. I used and tailored this because using it is faster than going through Adobe Acrobat and splitting files that way. A problem I had was that if I did two different documents, then the pages would overwrite each other, so I came up with a simple way to make sure that they all saved as individual files. This could use some work but it works for this purpose and I rename the files anyway so I don't want to spend too much time on it, or it defeats the purpose of making a program to save me time.