Add files via upload
I had to make hundreds, if not over a thousand subfolders. Every street has a segment with a unique ID, every street has one or more culverts. I had to make folders for every one of these culverts within the corresponding segment folders. This saved me a lot of time.